requirements/dev.txt
-r prod.txt
black==20.8b1
flake8==3.8.4
pytest==6.2.2
pytest-django==4.1.0
django-factory-boy==1.0.0
pytest-env==0.6.2
pytest-mock==3.5.1
